requirements.txt
Django>=3.2,<5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-filter==25.1
django-cors-headers==4.7.0
pre_commit==4.2.0
django-environ==0.11.2
python-dotenv==1.1.0
openpyxl==3.1.5
